Warnings as errors and further fixes

+++ src/Resources/Registries/Conveyors/ConveyorRegistry.cpp
@@ -40,7 +40,8 @@ cpp_conv::resources::ResourceAsset* conveyorAssetHandler(cpp_conv::resources::re
 {
     const auto pStrData = reinterpret_cast<const char*>(rData.m_pData);
 
-    const std::string copy(pStrData, rData.m_uiSize / sizeof(char));
+    // ReSharper disable once CppRedundantCastExpression
+    const std::string copy(pStrData, (int)(rData.m_uiSize / sizeof(char)));
     std::istringstream ss(copy);
 
     std::string id;
